test: split added transistor scenarios into one-test-per-file
